server/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-filter>=24.2
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-django>=4.9.0
gunicorn>=22.0.0
whitenoise>=6.7.0
